Create model classes and add file-based DB with web view
The model is constructed as follows:
users have dislike and like objects where we save the the ID's of the
like/dislike giver and taker in separate attributes. This structure can
be helpful when we ask the DB for likes in order to create matches and
dislikes in order to not show the disliked statuses in the match finding
game.

userStatuses are objects that belong to a certain user. This implementation
helps finding the owner of the status in case of like or dislike.

users have a list of matches and matches have messages which belong to users.