Login and register user
added endpoints /login and /register in the AuthenticationController to
test registering a user by writing the data in the DB and logging by searching
the DB for the user with the given Email and Password.

on registering there is also the status that is required to be given
and thus a UserStatus object is created with the given status and user.
UserStatus has its own repository.