Implement likes on statuses
when a use likes a status, a post request is sent on the endpoint
/filter/status/{id}/like and the a new entry is created in table of likes
with the id of the like giver and the id of the like taker which is
the owner of the status.

users cannot send requests to like own status and if a like with the
tuple is found in the like table, the like is not saved(as it exists).