Retrieve only statuses not like or disliked
as users should not see the statuses they liked or disliked in when
finding a match, we search the DB for all the status except ours, all the
statuses we liked, all the statuses we disliked and then substract
the latter 2 from the first. This way we get only the relevant statuses
we had no interaction with.